Modify max ACK length and idle time calculation

+++ os/net/mac/tsch/tsch-log.h
@@ -102,7 +102,7 @@ struct tsch_log_t {
       uint32_t app_seqno;
 #endif
       uint16_t unused_offset_time;
-      uint16_t idle_time;
+      int idle_time;
     } tx;
     struct {
       linkaddr_t  src;
@@ -120,7 +120,7 @@ struct tsch_log_t {
       uint32_t app_seqno;
 #endif
       uint16_t unused_offset_time;
-      uint16_t idle_time;
+      int idle_time;
     } rx;
   };
 };
